# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BookBridge – Student Book Exchange Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List books with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: semester
          in: query
          schema:
            type: string
        - name: condition
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: listingType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"

    post:
      operationId: createBook
      tags: [books]
      summary: Create a new book listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
...
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Categories
...
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  # Wishlist
...
  /admin/categories:
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [admin]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Book:
      type: object
      required: [id, title, author, publisher, category, department, semester, condition, price, listingType, isAvailable, sellerId, sellerName, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        category:
          type: string
        department:
          type: string
        semester:
          type: string
        condition:
          type: string
        price:
          type: number
        listingType:
          type: string
          enum: [sell, exchange, donate]
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        sellerId:
          type: integer
        sellerName:
          type: string
        viewCount:
          type: integer
        wishlistCount:
          type: integer
        createdAt:
          type: string
    BookDetail:
      type: object
      required: [id, title, author, publisher, category, department, semester, condition, price, listingType, isAvailable, sellerId, sellerName, sellerEmail, sellerPhone, sellerCollege, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        category:
          type: string
        department:
          type: string
        semester:
          type: string
        condition:
          type: string
        price:
          type: number
        listingType:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerEmail:
          type: string
        sellerPhone:
          type: ["string", "null"]
        sellerCollege:
          type: string
        viewCount:
          type: integer
        wishlistCount:
          type: integer
        isInWishlist:
          type: boolean
        createdAt:
          type: string
    BookInput:
      type: object
      required: [title, author, publisher, category, department, semester, condition, price, listingType]
      properties:
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        category:
          type: string
        department:
          type: string
        semester:
          type: string
        condition:
          type: string
          enum: [new, like-new, good, fair, poor]
        price:
          type: number
        listingType:
          type: string
          enum: [sell, exchange, donate]
        description:
          type: string
        imageUrl:
          type: string
    BookUpdate:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        category:
          type: string
        department:
          type: string
        semester:
          type: string
        condition:
          type: string
        price:
          type: number
        listingType:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        isAvailable:
          type: boolean
    BookListResponse:
      type: object
      required: [books, total, page, totalPages]
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer

    # Categories
    Category:
      type: object
      required: [id, name, icon, bookCount]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        bookCount:
          type: integer
    CategoryInput:
      type: object
      required: [name, icon]
      properties:
        name:
          type: string
        icon:
          type: string

    # Wishlist
...