# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BookBridge – Student Book Exchange Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookInputCondition } from './bookInputCondition';
import type { BookInputListingType } from './bookInputListingType';

export interface BookInput {
  title: string;
  author: string;
  publisher: string;
  category: string;
  department: string;
  semester: string;
  condition: BookInputCondition;
  price: number;
  listingType: BookInputListingType;
  description?: string;
  imageUrl?: string;
}